1. Бросил UserFriendlyException вместо Exception в Core сервисах 2. Если sum=0, сразу вовзращаю 0 из Core сервиса 3. Сделал статическую переменную класса Random. 4. Исправил опечатку currencyCode. 5. Поменял тип int на decimal. 6. Добавил .gitignore

+++ MiniBank.Core/Interfaces/ICurrencyCode.cs
@@ -2,6 +2,6 @@
 {
     public interface ICurrencyCode
     {
-        public int GetExchangeRate(string currencyCode);
+        public decimal GetExchangeRate(string currencyCode);
     }
 }
